Compilation fixed

+++ src/atrac/atrac_psy_common.cpp
@@ -5,13 +5,13 @@ namespace NAtracDEnc {
 //returns 1 for tone-like, 0 - noise-like
 TFloat AnalizeScaleFactorSpread(const std::vector<TScaledBlock>& scaledBlocks) {
     TFloat s = 0.0;
-    for (int i = 0; i < scaledBlocks.size(); ++i) {
+    for (size_t i = 0; i < scaledBlocks.size(); ++i) {
         s += scaledBlocks[i].ScaleFactorIndex;
     }
     s /= scaledBlocks.size();
     TFloat sigma = 0.0;
     TFloat t = 0.0;
-    for (int i = 0; i < scaledBlocks.size(); ++i) {
+    for (size_t i = 0; i < scaledBlocks.size(); ++i) {
         t = (scaledBlocks[i].ScaleFactorIndex - s);
         t *= t;
         sigma += t;
